Fix #22: Clean up elf2cfetbl build logic
Remove the now-unnecessary inclusion of CFE TBL internal definitions.
All logic in this tool is based on public include files only.

CFE ensures that the header sizes are always a multiple of 4 bytes,
so the logic to selectively byte-align is unnecessary.  CFE now
requires this to be 4-byte aligned.

Clean up all hacks in the CMake recipe, the build is now much
more straightforward when internal headers aren't being used.