Fix #48, string table identification.
Support ELF files that have all strings, including ELF section names,
in one single ".strtab" section in the ELF file.

This gives preferential treatment to a section named ".strtab" as this
should always be the section which contains the names for the ".symtab"
section.

This should also be true of ELF files that contain other strtab sections,
such as a ".dynstr" section.